\name{HTSdata}
\alias{HTSdata}
\docType{data}
\title{
HTSdata
}
\description{
Modified raw data from an HTS experiment
}
\usage{data(HTSdata)}
\format{
  A data frame with 7116 observations on the following 8 variables.
  \describe{
    \item{\code{SOBARCODE}}{a factor with levels \code{SO000001} \code{SO000002}}
    \item{\code{BARCODE}}{a factor with levels \code{PL000001} \code{PL000002} \code{PL000003} \code{PL000004} \code{PL000005} \code{PL000006} \code{PL000007} \code{PL000008} \code{PL000009} \code{PL000010} \code{PL000011} \code{PL000012}}
    \item{\code{XPOS}}{a numeric vector}
    \item{\code{YPOS}}{a numeric vector}
    \item{\code{WELL_USAGE}}{a factor with levels \code{mock1} \code{mock2} \code{negCTRL} \code{posCTRL1} \code{posCTRL2} \code{posCTRL3} \code{Sample}}
    \item{\code{Compound}}{a factor with levels \code{  -         -    } \code{Cmpd100001} \code{Cmpd100002} \code{Cmpd100003} \code{Cmpd100004} \code{Cmpd100005} \code{Cmpd100006} \code{Cmpd100007} \code{Cmpd100008} \code{Cmpd100009} \code{Cmpd100010} \code{Cmpd100011} \code{Cmpd100012} \code{Cmpd100013} \code{Cmpd100014} \code{Cmpd100015} \code{Cmpd100016} \code{Cmpd100017} \code{Cmpd100018} \code{Cmpd100019} \code{Cmpd100020} \code{Cmpd100021} \code{Cmpd100022} \code{Cmpd100023} \code{Cmpd100024} \code{Cmpd100025} \code{Cmpd100026} \code{Cmpd100027} \code{Cmpd100028} \code{Cmpd100029} \code{Cmpd100030} \code{Cmpd100031} \code{Cmpd100032} \code{Cmpd100033} \code{Cmpd100034} \code{Cmpd100035} \code{Cmpd100036} \code{Cmpd100037} \code{Cmpd100038} \code{Cmpd100039} \code{Cmpd100040} \code{Cmpd100041} \code{Cmpd100042} \code{Cmpd100043} \code{Cmpd100044} \code{Cmpd100045} \code{Cmpd100046} \code{Cmpd100047} \code{Cmpd100048} \code{Cmpd100049} \code{Cmpd100050} \code{Cmpd100051} \code{Cmpd100052} \code{Cmpd100053} \code{Cmpd100054} \code{Cmpd100055} \code{Cmpd100056} \code{Cmpd100057} \code{Cmpd100058} \code{Cmpd100059} \code{Cmpd100060} \code{Cmpd100061} \code{Cmpd100062} \code{Cmpd100063} \code{Cmpd100064} \code{Cmpd100065} \code{Cmpd100066} \code{Cmpd100067} \code{Cmpd100068} \code{Cmpd100069} \code{Cmpd100070} \code{Cmpd100071} \code{Cmpd100072} \code{Cmpd100073} \code{Cmpd100074} \code{Cmpd100075} \code{Cmpd100076} \code{Cmpd100077} \code{Cmpd100078} \code{Cmpd100079} \code{Cmpd100080} \code{Cmpd100081} \code{Cmpd100082} \code{Cmpd100083} \code{Cmpd100084} \code{Cmpd100085} \code{Cmpd100086} \code{Cmpd100087} \code{Cmpd100088} \code{Cmpd100089} \code{Cmpd100090} \code{Cmpd100091} \code{Cmpd100092} \code{Cmpd100093} \code{Cmpd100094} \code{Cmpd100095} \code{Cmpd100096} \code{Cmpd100097} \code{Cmpd100098} \code{Cmpd100099} \code{Cmpd100100} \code{Cmpd100101} \code{Cmpd100102} \code{Cmpd100103} \code{Cmpd100104} \code{Cmpd100105} \code{Cmpd100106} \code{Cmpd100107} \code{Cmpd100108} \code{Cmpd100109} \code{Cmpd100110} \code{Cmpd100111} \code{Cmpd100112} \code{Cmpd100113} \code{Cmpd100114} \code{Cmpd100115} \code{Cmpd100116} \code{Cmpd100117} \code{Cmpd100118} \code{Cmpd100119} \code{Cmpd100120} \code{Cmpd100121} \code{Cmpd100122} \code{Cmpd100123} \code{Cmpd100124} \code{Cmpd100125} \code{Cmpd100126} \code{Cmpd100127} \code{Cmpd100128} \code{Cmpd100129} \code{Cmpd100130} \code{Cmpd100131} \code{Cmpd100132} \code{Cmpd100133} \code{Cmpd100134} \code{Cmpd100135} \code{Cmpd100136} \code{Cmpd100137} \code{Cmpd100138} \code{Cmpd100139} \code{Cmpd100140} \code{Cmpd100141} \code{Cmpd100142} \code{Cmpd100143} \code{Cmpd100144} \code{Cmpd100145} \code{Cmpd100146} \code{Cmpd100147} \code{Cmpd100148} \code{Cmpd100149} \code{Cmpd100150} \code{Cmpd100151} \code{Cmpd100152} \code{Cmpd100153} \code{Cmpd100154} \code{Cmpd100155} \code{Cmpd100156} \code{Cmpd100157} \code{Cmpd100158} \code{Cmpd100159} \code{Cmpd100160} \code{Cmpd100161} \code{Cmpd100162} \code{Cmpd100163} \code{Cmpd100164} \code{Cmpd100165} \code{Cmpd100166} \code{Cmpd100167} \code{Cmpd100168} \code{Cmpd100169} \code{Cmpd100170} \code{Cmpd100171} \code{Cmpd100172} \code{Cmpd100173} \code{Cmpd100174} \code{Cmpd100175} \code{Cmpd100176} \code{Cmpd100177} \code{Cmpd100178} \code{Cmpd100179} \code{Cmpd100180} \code{Cmpd100181} \code{Cmpd100182} \code{Cmpd100183} \code{Cmpd100184} \code{Cmpd100185} \code{Cmpd100186} \code{Cmpd100187} \code{Cmpd100188} \code{Cmpd100189} \code{Cmpd100190} \code{Cmpd100191} \code{Cmpd100192} \code{Cmpd100193} \code{Cmpd100194} \code{Cmpd100195} \code{Cmpd100196} \code{Cmpd100197} \code{Cmpd100198} \code{Cmpd100199} \code{Cmpd100200} \code{Cmpd100201} \code{Cmpd100202} \code{Cmpd100203} \code{Cmpd100204} \code{Cmpd100205} \code{Cmpd100206} \code{Cmpd100207} \code{Cmpd100208} \code{Cmpd100209} \code{Cmpd100210} \code{Cmpd100211} \code{Cmpd100212} \code{Cmpd100213} \code{Cmpd100214} \code{Cmpd100215} \code{Cmpd100216} \code{Cmpd100217} \code{Cmpd100218} \code{Cmpd100219} \code{Cmpd100220} \code{Cmpd100221} \code{Cmpd100222} \code{Cmpd100223} \code{Cmpd100224} \code{Cmpd100225} \code{Cmpd100226} \code{Cmpd100227} \code{Cmpd100228} \code{Cmpd100229} \code{Cmpd100230} \code{Cmpd100231} \code{Cmpd100232} \code{Cmpd100233} \code{Cmpd100234} \code{Cmpd100235} \code{Cmpd100236} \code{Cmpd100237} \code{Cmpd100238} \code{Cmpd100239} \code{Cmpd100240} \code{Cmpd100241} \code{Cmpd100242} \code{Cmpd100243} \code{Cmpd100244} \code{Cmpd100245} \code{Cmpd100246} \code{Cmpd100247} \code{Cmpd100248} \code{Cmpd100249} \code{Cmpd100250} \code{Cmpd100251} \code{Cmpd100252} \code{Cmpd100253} \code{Cmpd100254} \code{Cmpd100255} \code{Cmpd100256} \code{Cmpd100257} \code{Cmpd100258} \code{Cmpd100259} \code{Cmpd100260} \code{Cmpd100261} \code{Cmpd100262} \code{Cmpd100263} \code{Cmpd100264} \code{Cmpd100265} \code{Cmpd100266} \code{Cmpd100267} \code{Cmpd100268} \code{Cmpd100269} \code{Cmpd100270} \code{Cmpd100271} \code{Cmpd100272} \code{Cmpd100273} \code{Cmpd100274} \code{Cmpd100275} \code{Cmpd100276} \code{Cmpd100277} \code{Cmpd100278} \code{Cmpd100279} \code{Cmpd100280} \code{Cmpd100281} \code{Cmpd100282} \code{Cmpd100283} \code{Cmpd100284} \code{Cmpd100285} \code{Cmpd100286} \code{Cmpd100287} \code{Cmpd100288} \code{Cmpd100289} \code{Cmpd100290} \code{Cmpd100291} \code{Cmpd100292} \code{Cmpd100293} \code{Cmpd100294} \code{Cmpd100295} \code{Cmpd100296} \code{Cmpd100297} \code{Cmpd100298} \code{Cmpd100299} \code{Cmpd100300} \code{Cmpd100301} \code{Cmpd100302} \code{Cmpd100303} \code{Cmpd100304} \code{Cmpd100305} \code{Cmpd100306} \code{Cmpd100307} \code{Cmpd100308} \code{Cmpd100309} \code{Cmpd100310} \code{Cmpd100311} \code{Cmpd100312} \code{Cmpd100313} \code{Cmpd100314} \code{Cmpd100315} \code{Cmpd100316} \code{Cmpd100317} \code{Cmpd100318} \code{Cmpd100319} \code{Cmpd100320} \code{Cmpd100321} \code{Cmpd100322} \code{Cmpd100323} \code{Cmpd100324} \code{Cmpd100325} \code{Cmpd100326} \code{Cmpd100327} \code{Cmpd100328} \code{Cmpd100329} \code{Cmpd100330} \code{Cmpd100331} \code{Cmpd100332} \code{Cmpd100333} \code{Cmpd100334} \code{Cmpd100335} \code{Cmpd100336} \code{Cmpd100337} \code{Cmpd100338} \code{Cmpd100339} \code{Cmpd100340} \code{Cmpd100341} \code{Cmpd100342} \code{Cmpd100343} \code{Cmpd100344} \code{Cmpd100345} \code{Cmpd100346} \code{Cmpd100347} \code{Cmpd100348} \code{Cmpd100349} \code{Cmpd100350} \code{Cmpd100351} \code{Cmpd100352} \code{Cmpd100353} \code{Cmpd100354} \code{Cmpd100355} \code{Cmpd100356} \code{Cmpd100357} \code{Cmpd100358} \code{Cmpd100359} \code{Cmpd100360} \code{Cmpd100361} \code{Cmpd100362} \code{Cmpd100363} \code{Cmpd100364} \code{Cmpd100365} \code{Cmpd100366} \code{Cmpd100367} \code{Cmpd100368} \code{Cmpd100369} \code{Cmpd100370} \code{Cmpd100371} \code{Cmpd100372} \code{Cmpd100373} \code{Cmpd100374} \code{Cmpd100375} \code{Cmpd100376} \code{Cmpd100377} \code{Cmpd100378} \code{Cmpd100379} \code{Cmpd100380} \code{Cmpd100381} \code{Cmpd100382} \code{Cmpd100383} \code{Cmpd100384} \code{Cmpd100385} \code{Cmpd100386} \code{Cmpd100387} \code{Cmpd100388} \code{Cmpd100389} \code{Cmpd100390} \code{Cmpd100391} \code{Cmpd100392} \code{Cmpd100393} \code{Cmpd100394} \code{Cmpd100395} \code{Cmpd100396} \code{Cmpd100397} \code{Cmpd100398} \code{Cmpd100399} \code{Cmpd100400} \code{Cmpd100401} \code{Cmpd100402} \code{Cmpd100403} \code{Cmpd100404} \code{Cmpd100405} \code{Cmpd100406} \code{Cmpd100407} \code{Cmpd100408} \code{Cmpd100409} \code{Cmpd100410} \code{Cmpd100411} \code{Cmpd100412} \code{Cmpd100413} \code{Cmpd100414} \code{Cmpd100415} \code{Cmpd100416} \code{Cmpd100417} \code{Cmpd100418} \code{Cmpd100419} \code{Cmpd100420} \code{Cmpd100421} \code{Cmpd100422} \code{Cmpd100423} \code{Cmpd100424} \code{Cmpd100425} \code{Cmpd100426} \code{Cmpd100427} \code{Cmpd100428} \code{Cmpd100429} \code{Cmpd100430} \code{Cmpd100431} \code{Cmpd100432} \code{Cmpd100433} \code{Cmpd100434} \code{Cmpd100435} \code{Cmpd100436} \code{Cmpd100437} \code{Cmpd100438} \code{Cmpd100439} \code{Cmpd100440} \code{Cmpd100441} \code{Cmpd100442} \code{Cmpd100443} \code{Cmpd100444} \code{Cmpd100445} \code{Cmpd100446} \code{Cmpd100447} \code{Cmpd100448} \code{Cmpd100449} \code{Cmpd100450} \code{Cmpd100451} \code{Cmpd100452} \code{Cmpd100453} \code{Cmpd100454} \code{Cmpd100455} \code{Cmpd100456} \code{Cmpd100457} \code{Cmpd100458} \code{Cmpd100459} \code{Cmpd100460} \code{Cmpd100461} \code{Cmpd100462} \code{Cmpd100463} \code{Cmpd100464} \code{Cmpd100465} \code{Cmpd100466} \code{Cmpd100467} \code{Cmpd100468} \code{Cmpd100469} \code{Cmpd100470} \code{Cmpd100471} \code{Cmpd100472} \code{Cmpd100473} \code{Cmpd100474} \code{Cmpd100475} \code{Cmpd100476} \code{Cmpd100477} \code{Cmpd100478} \code{Cmpd100479} \code{Cmpd100480} \code{Cmpd100481} \code{Cmpd100482} \code{Cmpd100483} \code{Cmpd100484} \code{Cmpd100485} \code{Cmpd100486} \code{Cmpd100487} \code{Cmpd100488} \code{Cmpd100489} \code{Cmpd100490} \code{Cmpd100491} \code{Cmpd100492} \code{Cmpd100493} \code{Cmpd100494} \code{Cmpd100495} \code{Cmpd100496} \code{Cmpd100497} \code{Cmpd100498} \code{Cmpd100499} \code{Cmpd100500} \code{Cmpd100501} \code{Cmpd100502} \code{Cmpd100503} \code{Cmpd100504} \code{Cmpd100505} \code{Cmpd100506} \code{Cmpd100507} \code{Cmpd100508} \code{Cmpd100509} \code{Cmpd100510} \code{Cmpd100511} \code{Cmpd100512} \code{Cmpd100513} \code{Cmpd100514} \code{Cmpd100515} \code{Cmpd100516} \code{Cmpd100517} \code{Cmpd100518} \code{Cmpd100519} \code{Cmpd100520} \code{Cmpd100521} \code{Cmpd100522} \code{Cmpd100523} \code{Cmpd100524} \code{Cmpd100525} \code{Cmpd100526} \code{Cmpd100527} \code{Cmpd100528} \code{Cmpd100529} \code{Cmpd100530} \code{Cmpd100531} \code{Cmpd100532} \code{Cmpd100533} \code{Cmpd100534} \code{Cmpd100535} \code{Cmpd100536} \code{Cmpd100537} \code{Cmpd100538} \code{Cmpd100539} \code{Cmpd100540} \code{Cmpd100541} \code{Cmpd100542} \code{Cmpd100543} \code{Cmpd100544} \code{Cmpd100545} \code{Cmpd100546} \code{Cmpd100547} \code{Cmpd100548} \code{Cmpd100549} \code{Cmpd100550} \code{Cmpd100551} \code{Cmpd100552} \code{Cmpd100553} \code{Cmpd100554} \code{Cmpd100555} \code{Cmpd100556} \code{Cmpd100557} \code{Cmpd100558} \code{Cmpd100559} \code{Cmpd100560} \code{Cmpd100561} \code{Cmpd100562} \code{Cmpd100563} \code{Cmpd100564} \code{Cmpd100565} \code{Cmpd100566} \code{Cmpd100567} \code{Cmpd100568} \code{Cmpd100569} \code{Cmpd100570} \code{Cmpd100571} \code{Cmpd100572} \code{Cmpd100573} \code{Cmpd100574} \code{Cmpd100575} \code{Cmpd100576} \code{Cmpd100577} \code{Cmpd100578} \code{Cmpd100579} \code{Cmpd100580} \code{Cmpd100581} \code{Cmpd100582} \code{Cmpd100583} \code{Cmpd100584} \code{Cmpd100585} \code{Cmpd100586} \code{Cmpd100587} \code{Cmpd100588} \code{Cmpd100589} \code{Cmpd100590} \code{Cmpd100591} \code{Cmpd100592} \code{Cmpd100593} \code{Cmpd100594} \code{Cmpd100595} \code{Cmpd100596} \code{Cmpd100597} \code{Cmpd100598} \code{Cmpd100599} \code{Cmpd100600} \code{Cmpd100601} \code{Cmpd100602} \code{Cmpd100603} \code{Cmpd100604} \code{Cmpd100605} \code{Cmpd100606} \code{Cmpd100607} \code{Cmpd100608} \code{Cmpd100609} \code{Cmpd100610} \code{Cmpd100611} \code{Cmpd100612} \code{Cmpd100613} \code{Cmpd100614} \code{Cmpd100615} \code{Cmpd100616} \code{Cmpd100617} \code{Cmpd100618} \code{Cmpd100619} \code{Cmpd100620} \code{Cmpd100621} \code{Cmpd100622} \code{Cmpd100623} \code{Cmpd100624} \code{Cmpd100625} \code{Cmpd100626} \code{Cmpd100627} \code{Cmpd100628} \code{Cmpd100629} \code{Cmpd100630} \code{Cmpd100631} \code{Cmpd100632}}
    \item{\code{Intensity}}{a numeric vector}
    \item{\code{log2Intensity}}{a numeric vector}
  }
}
\details{
Modified raw data from an HTS experiment for demonstrating the utility
of our package displayHTS 
}
\examples{
data(HTSdata)
boxplot(HTSdata[,"log2Intensity"] ~ HTSdata[, "WELL_USAGE"], cex.axis=0.75)
}
\keyword{datasets}
